refactor: problems panel details
