--- src/features/diagramWorkspace/types.ts
@@ -1,5 +1,0 @@
-import type { ReactNode } from "react";
-
-export type DiagramWorkspaceProps = {
-  diagram?: ReactNode;
-};
